restructured paths
